# docs/contracts/openapi/current/public-openapi.json
{
  "components": {
    "schemas": {
...
      "EventIngestIn": {
        "additionalProperties": false,
        "description": "Anonymous aggregate event payload (§I07).\n\nStrict privacy policy:\nextra = 'forbid' rejects any unknown fields (such as email, visitor_id,\ncookie, tokens, full referrers).",
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "pagePath": {
            "title": "Pagepath",
            "type": "string"
          },
          "target": {
            "default": "",
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "event",
          "pagePath",
          "locale"
        ],
        "title": "EventIngestIn",
        "type": "object"
      },
      "EventIngestOut": {
        "description": "Ingest acceptance response.",
        "properties": {
          "status": {
            "default": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "EventIngestOut",
        "type": "object"
      },
...
      "PagedBookListOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookListOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedBookListOut",
        "type": "object"
      },
...
      "PagedProjectListOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectListOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedProjectListOut",
        "type": "object"
      },
      "PagedPublicationListOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicationListOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedPublicationListOut",
        "type": "object"
      },
...
      "PagedTalkListOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TalkListOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedTalkListOut",
        "type": "object"
      },
...
      "ResearchStatementOut": {
        "description": "Public research statement (sanitized rich text + optional active PDF).",
        "properties": {
          "alternates": {
            "items": {
              "$ref": "#/components/schemas/AlternateLocaleOut"
            },
            "title": "Alternates",
            "type": "array"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "relatedRecords": {
            "items": {
              "$ref": "#/components/schemas/WorkRefOut"
            },
            "title": "Relatedrecords",
            "type": "array"
          },
          "seo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicSeoOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "statement_pdf": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicMediaOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryDocumentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "locale",
          "slug",
          "title",
          "body",
          "published_at",
          "updated_at"
        ],
        "title": "ResearchStatementOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "title": "Taha CMS Public API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_books",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedBookListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published books for a locale (paginated)"
      }
    },
...
    "/api/projects/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_projects",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "has_case_study",
            "required": false,
            "schema": {
              "default": false,
              "title": "Has Case Study",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedProjectListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published projects shown on /projects/ (paginated)"
      }
    },
...
    "/api/publications/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_publications",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedPublicationListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published publications for a locale (paginated)"
      }
    },
...
    "/api/research/projects/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_research_projects",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedProjectListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published projects for a locale (paginated)"
      }
    },
...
    "/api/research/publications/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_research_publications",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedPublicationListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published publications for a locale (paginated)"
      }
    },
...
    "/api/research/statements/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_research_statements",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResearchStatementOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published research statements for a locale"
      }
    },
...
    "/api/talks/{locale}": {
      "get": {
        "operationId": "apps_api_api_list_talks",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedTalkListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published talks for a locale (paginated)"
      }
    },
...
    "/api/v1/analytics/events": {
      "post": {
        "description": "Ingest anonymous first-party event and update daily aggregate counters.\n\nA09 validation (§I07, fail-closed, I08 envelope on every rejection):\n\n- Oversized bodies are rejected before anything is stored.\n- Foreign cross-origin posts are rejected.\n- Per-IP rate limits are enforced.\n- ``event`` must be a known enum member; ``locale`` must be fa/en.\n- ``pagePath`` must be a canonical route whose locale matches ``locale``\n  (the locale-neutral gateway ``/`` is accepted for either locale).\n- ``target`` must be a registered action ID for the event (page views\n  carry no target; only ``cv_download`` has shipped sub-targets).\n\nPrivacy & Security: persists zero visitor identifiers; unknown fields are\nrejected by the schema (422 envelope); failed analytics never block\nnavigation or forms.",
        "operationId": "apps_analytics_api_ingest_event",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIngestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventIngestOut"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventIngestOut"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeOut"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeOut"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Ingest first-party anonymous aggregate event"
      }
    },
    "/api/v1/collections/{locale}": {
      "get": {
        "description": "List published collections with pagination (1-based, default 20, max 50).",
        "operationId": "apps_api_api_list_collections",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published collections for a locale (paginated)"
      }
    },
...
    "/api/v1/lessons/{locale}": {
      "get": {
        "description": "List published lessons for a course ordered by position, id.\n\nGuard: Both parent course and lesson must be published in the requested locale.\nIf parent course is not published or does not exist, returns count 0.",
        "operationId": "apps_api_api_list_lessons",
        "parameters": [
          {
            "in": "path",
            "name": "locale",
            "required": true,
            "schema": {
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "course",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonListOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List published course lessons for a locale"
      }
    },
...
  },
  "servers": []
}